Added exploration theme and audio manager script
Exploration theme has a baseline synth playing, with random variation in what melody it plays, and which random background noises.

Exploration theme now plays when player is outside of hub.

Hub theme now plays when player is inside of hub.

Also created an audio manager script with the help of YouTube (thank god that exists). The audio manager script is where all audio controlling is handled.